moved add and update community to features
