--- src/components/community/UpdateCommunity.js
@@ -1,152 +1,0 @@
-import { TextField,FormControl, Select, MenuItem, InputLabel, 
-         Button ,FormControlLabel, Checkbox , Grid, Box} from "@mui/material";
-import Logo from "layout/Logo";
-import PageContainer from "layout/PageContainer";
-import PageTitle from "layout/PageTitle";
-import axios from "axios";
-import { useState, useEffect } from "react";
-
-
-const UpdateCommunity = (props) => {
-const [communityName, setCommunityName] = useState('')
-const [communityManager, setCommunityManager] = useState('')
-const [isActive, setIsActive] = useState(true)
-const [communityDescription, setCommunityDescription] = useState('')
-const [communityManagers , setCommunityManagers] = useState([])
-
-
-useEffect(() => {
-    let isMounted = true;
-      
-    axios({
-        method: "GET",
-        url: 'http://localhost:8000/api/managers'
-
-    }).then(res => {
-      if (isMounted) setCommunityManagers(res.data.data); 
-    })
-    return () => { isMounted = false }; 
-  },[]); 
-
-
-const handleCommunityNameChange = (e) => {
-    setCommunityName(e.target.value)
-}
-
-const handleManagerChange = (e) => {
-    setCommunityManager(e.target.value)
-}
-
-const handleActiveChange = (e) => {
-    setIsActive(!isActive)
-}
-
-const handleDescriptionChange = (e) => {
-    setCommunityDescription(e.target.value)
-}
-
-const handleOnSave = (e) => {
-
-    const data = {
-        Communityname: communityName,
-        Communitymgrpeopleid: communityManager,
-        Isactive: isActive,
-        Communitydescription: communityDescription
-    }
-
-        /*
-        axios({
-            method: "POST",
-            url: 'http://localhost:8000/api/community',
-            data: data
-
-        })
-        */
-
-    console.log(data)
-}
-
-return ( 
-    <PageContainer>
-    <Logo/>
-    <PageTitle title="Community Update Page"/>
-        <Grid container >
-            <Grid item xs={12} sm={5}>
-                <TextField fullWidth value={communityName} onChange={handleCommunityNameChange} id="communityName" 
-                    label="Name of Community" 
-                    variant="outlined" 
-                    sx={{
-                        mt: 5,
-                        backgroundColor: '#FFFFFF'
-                    }}/>
-            </Grid>
-            <Grid item xs={0} sm={7}>
-            </Grid>
-
-            <Grid item xs={12} sm={5}>
-                <FormControl sx={{
-                    mt: 5
-                }} fullWidth>
-                <InputLabel id="communityAssigned">Community Assigned To</InputLabel>
-                <Select
-                    
-                    id="communityManager"
-                    value={communityManager}
-                    label="Community Assigned To"
-                    onChange={handleManagerChange}
-                    sx={{
-                        backgroundColor: "#FFFFFF"
-                    }}
-                >
-                    {communityManagers.map((manager) => {
-                        return(
-                            <MenuItem key={manager.people_id} value={manager.people_id}> {manager.first_name + " " + manager.last_name} </MenuItem>
-                        )
-                    })}
-
-                </Select>
-                </FormControl>                    
-            </Grid>
-            <Grid item xs={12} sm={7}>
-                <Box sx={{
-                    display :"flex",
-                    mt: 5,
-                    pt: 3,
-                    justifyContent: "center",
-                    
-                }}>
-                    <FormControlLabel control={<Checkbox checked={isActive} onChange={handleActiveChange}/>} label="Active" align='center' />
-                </Box>
-
-            </Grid>
-            <Grid item xs={12} sm={7}>
-            <TextField
-                fullWidth
-                id="communityDesc"
-                label="Community Description"
-                multiline
-                rows={6}
-                sx={{
-                    backgroundColor: "#FFFFFF",
-                    mt:5
-                }}
-                value={communityDescription}
-                onChange={handleDescriptionChange}
-            />                        
-            </Grid>
-            <Grid item sm={5}/>
-            <Grid item sm={8}/>
-            <Grid item xs={12} sm={4}>
-                <Button variant="contained" sx={{
-                        mt: 5,
-                        mb: 5,
-                        width: "100%"
-                        }} onClick={handleOnSave}>Update</Button>
-
-            </Grid>
-        </Grid>
-</PageContainer>
-);
-}
-
-export default UpdateCommunity;
