+++ src/features/index.js
@@ -1,3 +1,5 @@
 import Communities from "./communities/";
+import AddCommunity from "./communities/AddCommunity"
+import  UpdateCommunity  from "./communities/UpdateCommunity";
 
-export { Communities };
+export { Communities , AddCommunity , UpdateCommunity};
