+++ src/index.js
@@ -3,6 +3,9 @@ import ReactDOM from 'react-dom/client';
 import 'index.css';
 import App from 'App';
 import reportWebVitals from 'reportWebVitals';
+import { AddCommunity } from 'features';
+import UpdateCommunity from 'features/communities/UpdateCommunity';
+
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
